Blogs/Blog_Api/requirements.txt
asgiref==3.11.0
Django==6.0
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
PyJWT==2.10.1
python-dotenv==1.2.1
sqlparse==0.5.4
